Fix for QC_only

+++ modules/local/quality_report.nf
@@ -21,8 +21,8 @@ process QUALITY_REPORT {
   """
 
   # Rename input files to published versions
-  mv $sample_coverage sample_coverage.txt
-  mv $amplicon_coverage amplicon_coverage.txt
+  test -f sample_coverage.txt || mv $sample_coverage sample_coverage.txt
+  test -f amplicon_coverage.txt || mv $amplicon_coverage amplicon_coverage.txt
 
   test -d quality_report || mkdir quality_report
   Rscript ${projectDir}/bin/cutadapt_summaryplots.R amplicon_coverage.txt sample_coverage.txt $amplicon_info quality_report
